pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "binance-futures-connector>=4.1.0",
    "ccxt>=4.5.22",
    "flask>=3.1.2",
    "pandas>=2.3.3",
]
